+++ src/emotional_state.py
