+++ src/personality.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 
-from core.emotional_state import EmotionalState
-from core.ocean_profile import OceanProfile
+from src.emotional_state import EmotionalState
+from src.ocean_profile import OceanProfile
 
 
 @dataclass
